requirements.txt
Flask==2.2.5
flask_apscheduler==1.13.1
Flask_Login==0.6.3
Flask_SQLAlchemy==2.5.1
flask_wtf==1.2.2
psutil==7.0.0
pytest==7.1.2
SQLAlchemy==1.4.49
python-dotenv==1.0.1
Flask-Talisman==1.1.0
dnspython
